# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/gamification_badges:
    get:
      tags:
        - GamificationBadges
      summary: List GamificationBadges
      description: Get a paginated list of GamificationBadges with optional filtering and sorting
      operationId: listGamificationBadges
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamificationBadgeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - GamificationBadges
      summary: Create GamificationBadge
      description: Create a new GamificationBadge
      operationId: createGamificationBadge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGamificationBadgeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamificationBadge'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/engagement_ratings:
    get:
      tags:
        - EngagementRatings
      summary: List EngagementRatings
      description: Get a paginated list of EngagementRatings with optional filtering and sorting
      operationId: listEngagementRatings
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementRatingList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - EngagementRatings
      summary: Create EngagementRating
      description: Create a new EngagementRating
      operationId: createEngagementRating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEngagementRatingRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementRating'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    GamificationBadge:
      type: object
      required:
        - id
        - name
        - active
        - rule_auth
        - rule_auth_user_ids
        - rule_auth_badge_ids
        - rule_max
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        active:
          type: boolean
        level:
          nullable: true
          type: string
        description:
          nullable: true
          type: string
        rule_auth:
          type: string
        rule_auth_user_ids:
          type: object
        rule_auth_badge_ids:
          type: object
        rule_max:
          type: boolean
        rule_max_number:
          nullable: true
          type: integer
          format: int32
        metadata:
          type: object
...
    CreateGamificationBadgeRequest:
      type: object
      required:
        - name
        - active
        - rule_auth
        - rule_auth_user_ids
        - rule_auth_badge_ids
        - rule_max
        - metadata
      properties:
        name:
          type: string
          maxLength: 200
        active:
          type: boolean
        level:
          nullable: true
          type: string
        description:
          nullable: true
          type: string
        rule_auth:
          type: string
        rule_auth_user_ids:
          type: object
        rule_auth_badge_ids:
          type: object
        rule_max:
          type: boolean
        rule_max_number:
          nullable: true
          type: integer
          format: int32
        metadata:
          type: object
...
    GamificationBadgeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GamificationBadge'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    EngagementRating:
      type: object
      required:
        - id
        - rated_model
        - rated_record_id
        - consumed
        - token_nonce
        - token_expires_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        rated_model:
          type: string
          maxLength: 120
        rated_record_id:
          type: string
          format: uuid
        parent_rated_model:
          nullable: true
          type: string
          maxLength: 120
        parent_rated_record_id:
          nullable: true
          type: string
          format: uuid
        rated_user_id:
          nullable: true
          type: string
          format: uuid
        rater_user_id:
          nullable: true
          type: string
          format: uuid
        rater_email:
          nullable: true
          type: string
          maxLength: 320
        rating_value:
          nullable: true
          type: integer
          format: int32
        feedback:
          nullable: true
          type: string
        consumed:
          type: boolean
        rated_on:
          nullable: true
          type: string
          format: date-time
        token_nonce:
          type: string
          maxLength: 64
        token_expires_at:
          type: string
          format: date-time
        publisher_comment:
          nullable: true
          type: string
        publisher_user_id:
          nullable: true
          type: string
          format: uuid
        publisher_replied_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateEngagementRatingRequest:
      type: object
      required:
        - rated_model
        - rated_record_id
        - consumed
        - token_nonce
        - token_expires_at
        - metadata
      properties:
        rated_model:
          type: string
          maxLength: 120
        rated_record_id:
          type: string
          format: uuid
        parent_rated_model:
          nullable: true
          type: string
          maxLength: 120
        parent_rated_record_id:
          nullable: true
          type: string
          format: uuid
        rated_user_id:
          nullable: true
          type: string
          format: uuid
        rater_user_id:
          nullable: true
          type: string
          format: uuid
        rater_email:
          nullable: true
          type: string
          maxLength: 320
        rating_value:
          nullable: true
          type: integer
          format: int32
        feedback:
          nullable: true
          type: string
        consumed:
          type: boolean
        rated_on:
          nullable: true
          type: string
          format: date-time
        token_nonce:
          type: string
          maxLength: 64
        token_expires_at:
          type: string
          format: date-time
        publisher_comment:
          nullable: true
          type: string
        publisher_user_id:
          nullable: true
          type: string
          format: uuid
        publisher_replied_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    EngagementRatingList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EngagementRating'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []